refactor(cost-sharing): split oversized validation files
